requirements.txt
flake8==7.2.0
pytest==8.3.5
black==25.1.0
isort==6.0.1
crewai
crewai-tools
python-dotenv
streamlit
openai

